Funkcja ruchu konia, wieży i dodanie paru niezbędnych rzeczy
